TS tests run with ts-node
